+++ src/parametric.rs
@@ -276,6 +276,7 @@ mod tests {
             name: name.to_owned(),
             inputs: Vec::new(),
             match_mode: mode,
+            tags: Vec::new(),
             on_hit: None,
         }
     }